requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.3.0
attrs==23.2.0
av==12.3.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==43.0.0
cycler==0.12.1
dnspython==2.6.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
idna==3.7
ifaddr==0.2.0
importlib-resources==6.4.0
intel-openmp==2021.4.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pyee==11.1.0
pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
streamlit-webrtc==0.47.7
sympy==1.13.1
tbb==2021.13.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.64
ultralytics-thop==2.0.0
urllib3==2.2.2
watchdog==4.0.1
zipp==3.19.2
